大型関数分離
