--- packages/URLFinder/src/debug_update.md
@@ -1,70 +1,0 @@
-# /contact/リンク検出問題 - デバッグ強化実装完了
-
-## 実装日時
-2025-01-07 09:30
-
-## 問題症状
-- 実在する `<a href="/contact/" class="l-header-gnav-contact-btn">Contact</a>` が検出されない
-- ログでは `"強み・会社情報" -> /about/` から処理開始（/contact/リンクが見つからない）
-
-## 根本原因分析
-1. **ナビゲーション要素は検出されている** (`Found 3 matches for navigation selector`)
-2. **しかし実際の /contact/ リンクが extractContactLinks で見つかっていない**
-
-## 改修実装内容
-
-### 1. ナビゲーションセレクター拡張
-```typescript
-// alleyoop.co.jp 専用セレクター追加
-/<[^>]*(?:class|id)=['"]*[^'"]*(?:l-header|header-gnav|gnav)[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-// Contact専用ボタンクラス  
-/<[^>]*class=['"]*[^'"]*(?:contact-btn|gnav-contact)[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi
-```
-
-### 2. ナビゲーション検索の詳細化
-- **各セレクターのマッチ数表示**: `Navigation selector ${i+1}: Found ${matches.length} matches`
-- **マッチ内容プレビュー**: `Analyzing navigation match ${j+1} (${match.length} chars): ${match.substring(0, 100)}...`
-- **全候補の収集**: 最高スコア候補を選択する方式に変更
-
-### 3. 強化されたデバッグログ
-```typescript
-// Contact専用検出ログ
-if (url.includes('/contact') || cleanLinkText.toLowerCase().includes('contact')) {
-  console.log(`🎯 CONTACT LINK DETECTED: "${cleanLinkText}" -> ${url}`);
-}
-
-// 全リンク候補表示（負スコアも含む）
-console.log(`Link candidate: "${cleanLinkText}" -> ${url} (score: ${totalScore}, reasons: ${allReasons.join(',')})`)
-
-// 結果の視覚的区別
-console.log(`✓ Contact link candidate: ...`)   // 正スコア
-console.log(`✗ Link excluded: ...`)            // 負スコア
-```
-
-### 4. TypeScript型安全性修正
-- navigationSelectors配列の型チェック強化
-- undefined値のガード条件追加
-
-## 期待される診断結果
-
-次回テスト実行で以下が確認できるはず：
-
-1. **ナビゲーション解析詳細**
-   - どのセレクターが `/contact/` リンクを含むHTML要素を検出するか
-   - 各マッチ内容のプレビュー表示
-
-2. **Contact検出ログ** 
-   - 🎯マークで `/contact/` リンクが実際に見つかるか表示
-   - `cleanLinkText` が正しく `"Contactお問い合わせ"` になるか
-
-3. **スコア計算詳細**
-   - なぜ `/contact/` が除外されるのか（負スコアの理由）
-   - `calculateContactPurity` の詳細な判定結果
-
-## 実装ステータス
-- ✅ TypeScriptエラー修正完了
-- ✅ デバッグログ強化完了  
-- ⏳ Google Apps Script環境でのテスト実行待ち
-
-## 次のアクション
-Google Apps Script環境でテスト実行し、デバッグログで根本原因を特定する。
