--- packages/URLFinder/src/implementation_complete.md
@@ -1,75 +1,0 @@
-# 200URL評価によるフォールバック改修 - 実装完了
-
-## 実装日時
-2025-01-07 10:00
-
-## 実装内容
-
-### 1. メイン処理で200URLリスト保存 ✅
-- **新規追加**: `validUrls: Array<{ url: string, pattern: string }>`配列
-- **保存タイミング**: メイン処理で200OK+有効ページが見つかった時点
-- **リセット処理**: `resetCandidates()`で初期化
-
-```typescript
-// 200 OK URLを記録（フォールバック用）
-this.validUrls.push({ url: testUrl, pattern: pattern });
-```
-
-### 2. フォールバック処理の簡素化 ✅
-- **削除対象**: Footer, Sidebar, MobileMenu, GeneralLinks, UrlPatternGuess
-- **残存機能**: Navigationのみの解析
-- **処理フロー**: Navigation → 200URL評価 → Not Found
-
-```typescript
-// Navigation search only
-const navResult = this.searchInNavigation(html, baseUrl);
-if (navResult.url && navResult.score > 0) {
-  return { contactUrl: navResult.url, ... };
-}
-
-// Stage 2: Evaluate 200 OK URLs
-const validUrlResult = this.evaluateValidUrls(baseUrl);
-```
-
-### 3. 200URL評価機能 ✅
-
-#### evaluateValidUrls関数
-- **評価順序**: HIGH_PRIORITY_PATTERNS順（/contact/ → /contact → ...）
-- **処理**: HTML再取得 → キーワード検出 → 閾値判定
-- **閾値**: 3種類以上のキーワードで問い合わせページと判定
-
-#### detectContactKeywords関数  
-- **対象**: HIGH_PRIORITY_CONTACT_KEYWORDS
-- **方式**: 大文字小文字区別なし、重複カウントなし
-- **戻り値**: `{ matchCount: number, foundKeywords: string[] }`
-
-### 4. ログ出力強化 ✅
-```typescript
-console.log(`=== Evaluating ${this.validUrls.length} valid URLs with keyword detection ===`);
-console.log(`✅ Contact page confirmed at ${urlInfo.url} (${keywordResult.matchCount} keywords: ...)`);
-console.log(`❌ Insufficient keywords at ${urlInfo.url} (${keywordResult.matchCount}/3 required)`);
-```
-
-## 期待される効果
-
-### alleyoop.co.jpでの動作予想
-1. **メイン処理**: `/contact/`が200OKで`validUrls`に保存
-2. **フォールバック - Navigation**: 解析結果が見つからない
-3. **フォールバック - 200URL評価**: 
-   - `/contact/` HTML再取得
-   - キーワード検出: "contact", "お問い合わせ", "Contact" 等
-   - 3種類以上 → ✅ 問い合わせページ確定
-
-### 処理効率の改善
-- **不要な段階削除**: Footer/Sidebar等の重い解析をスキップ
-- **確実な結果**: 実在する200URLのみを評価対象
-
-## 実装状況
-- ✅ 200URLリスト保存機能
-- ✅ フォールバック処理簡素化  
-- ✅ キーワード検出機能
-- ✅ 統合ログ出力
-- ⏳ Google Apps Script環境でのテスト実行待ち
-
-## 次のアクション
-Google Apps Script環境でテスト実行し、alleyoop.co.jpで正しく`/contact/`が検出されることを確認する。
