--- packages/URLFinder/src/improvement_log.md
@@ -1,129 +1,0 @@
-# URLFinder 包括的検索機能 実装完了レポート
-
-## 実装日時
-2025-01-06 23:25
-
-## 改修概要
-alleyoop.co.jpで実在する/contact/リンクが検出されない問題を解決するため、HTML解析の包括的検索機能を実装。
-
-## 改修内容
-
-### 1. HTML検索段階の拡張（6段階に拡張）
-- **Stage 1**: Navigation search (既存・拡張済み)  
-- **Stage 2**: Footer search (既存・拡張済み)
-- **Stage 3**: **Sidebar search (新規追加・実装完了)** ✅
-- **Stage 4**: **Mobile menu search (新規追加・実装完了)** ✅  
-- **Stage 5**: General links search (既存)
-- **Stage 6**: URL pattern guessing (既存)
-
-### 2. 新規実装関数
-
-#### `searchInSidebar(html: string, baseUrl: string)`
-```typescript
-// サイドバー専用セレクター
-const sidebarSelectors = [
-  /<aside[\s\S]*?<\/aside>/gi,
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:sidebar|side-menu|side-nav)[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-  /<div[^>]*(?:class|id)=['"]*[^'"]*(?:widget|sidebar)[^'"]*['"][^>]*>[\s\S]*?<\/div>/gi
-];
-```
-
-#### `searchInMobileMenu(html: string, baseUrl: string)`
-```typescript
-// モバイルメニュー専用セレクター  
-const mobileSelectors = [
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:mobile-menu|hamburger|drawer)[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:toggle|collapse|accordion)[^'"]*menu[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:responsive|mobile)[^'"]*nav[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi
-];
-```
-
-### 3. リンクテキストパターンの大幅拡張 ✅
-
-#### HIGH_PRIORITY_CONTACT_KEYWORDS拡張
-**英語パターン追加:**
-- 'contact form', 'enquiry', 'get in touch', 'reach out'  
-- 'send message', 'message us'
-
-**日本語パターン追加:**
-- 'お問い合わせフォーム', 'お問い合わせはこちら', '問い合わせフォーム'
-
-#### MEDIUM_PRIORITY_CONTACT_KEYWORDS拡張
-**新規追加:**
-- 'submit', 'send', 'mail form', 'feedback', 'quote request'
-- 'お見積もり', '資料請求'
-
-### 4. 検索ロジックの統合と優先順位制御
-
-#### 早期終了条件
-- **HIGH_CONFIDENCE_THRESHOLD**: 15点以上で即座に結果確定
-- **MEDIUM_CONFIDENCE_THRESHOLD**: 8点以上で条件付き確定
-- **MINIMUM_ACCEPTABLE_THRESHOLD**: 3点以上で最低要件クリア
-
-#### コンテキスト別ボーナス
-- **navigation**: +5点（既存）
-- **footer**: +3点（既存）  
-- **sidebar**: +2点（新規）
-- **mobile_menu**: +4点（新規）
-
-### 5. ナビゲーション検索の拡張（既存改良）
-
-#### navigationSelectors拡張
-```typescript
-const navigationSelectors = [
-  /<nav[\s\S]*?<\/nav>/gi,
-  /<header[\s\S]*?<\/header>/gi,
-  // メインメニュー系のクラス/ID
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:nav|menu|navigation|main-menu|header-menu)[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-  // モバイルメニュー系（ナビゲーション内）
-  /<[^>]*(?:class|id)=['"]*[^'"]*(?:mobile|hamburger|toggle|responsive)[^'"]*menu[^'"]*['"][^>]*>[\s\S]*?<\/[^>]+>/gi,
-  // ul/liベースのメニュー
-  /<ul[^>]*(?:class|id)=['"]*[^'"]*(?:nav|menu)[^'"]*['"][^>]*>[\s\S]*?<\/ul>/gi
-];
-```
-
-## 期待される改善効果
-
-### 1. リンク検出率の向上
-- **従来**: ナビゲーション・フッターのみ検索
-- **改修後**: ナビ・フッター・サイドバー・モバイルメニューを包括的に検索
-
-### 2. キーワードマッチング精度向上  
-- **従来**: 基本的な contact/お問い合わせ のみ
-- **改修後**: 英語・日本語の多様な表現に対応
-
-### 3. 段階的早期終了による高効率化
-- 高信頼度リンクを即座に確定
-- 不要な検索処理をスキップして実行時間短縮
-
-## alleyoop.co.jp での期待される結果
-
-### 問題の症状
-- 実在する `/contact/` リンクが検出されない
-- 代わりに `/service/strategy/` が選択される問題
-
-### 改修による解決
-1. **拡張されたナビゲーション検索**でヘッダー内の contact リンクを検出
-2. **フッター検索**でページ下部の contact リンクを検出  
-3. **モバイルメニュー検索**でハンバーガーメニュー内の contact を検出
-4. **拡張キーワード**により 'Contact', 'Contact us' 等の英語表記に対応
-
-### 実行推奨
-```javascript
-// テスト実行
-function main() {
-  const testUrl = 'https://www.alleyoop.co.jp/';
-  const result = findContactPage(testUrl);
-  console.log('Contact URL:', result.contactUrl);
-  console.log('Search Method:', result.searchMethod);  
-  console.log('Found Keywords:', result.foundKeywords);
-}
-```
-
-## 実装ステータス
-- ✅ HTML検索段階拡張 (Stage 3, 4 追加)
-- ✅ リンクテキストパターン拡張
-- ✅ 検索ロジック統合と早期終了実装
-- ✅ analyzeHtmlContent関数への統合完了
-
-**次のステップ**: Google Apps Script環境でのテスト実行とログ確認
