--- packages/URLFinder/src/progress.html
@@ -1,275 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <style>
-    body {
-      font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-      margin: 0;
-      padding: 20px;
-      background-color: #f8f9fa;
-    }
-    
-    .container {
-      background: white;
-      border-radius: 8px;
-      padding: 24px;
-      box-shadow: 0 2px 8px rgba(0,0,0,0.1);
-      max-width: 360px;
-      margin: 0 auto;
-      text-align: center;
-    }
-    
-    .title {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 18px;
-      font-weight: 600;
-      margin-bottom: 20px;
-      color: #1a73e8;
-    }
-    
-    .title .icon {
-      margin-right: 8px;
-      font-size: 20px;
-    }
-    
-    .message {
-      color: #5f6368;
-      margin-bottom: 24px;
-      font-size: 14px;
-    }
-    
-    .progress-container {
-      margin: 20px 0;
-    }
-    
-    .progress-bar {
-      width: 100%;
-      height: 8px;
-      background-color: #e8eaed;
-      border-radius: 4px;
-      overflow: hidden;
-      margin-bottom: 12px;
-    }
-    
-    .progress-fill {
-      height: 100%;
-      background: linear-gradient(90deg, #1a73e8, #4285f4);
-      border-radius: 4px;
-      transition: width 0.3s ease;
-      animation: pulse 2s infinite;
-    }
-    
-    @keyframes pulse {
-      0% { opacity: 0.6; }
-      50% { opacity: 1; }
-      100% { opacity: 0.6; }
-    }
-    
-    .progress-text {
-      font-size: 14px;
-      color: #5f6368;
-      margin-bottom: 16px;
-    }
-    
-    .stats {
-      display: flex;
-      justify-content: space-around;
-      margin: 16px 0;
-      font-size: 13px;
-    }
-    
-    .stat-item {
-      text-align: center;
-    }
-    
-    .stat-value {
-      display: block;
-      font-weight: 600;
-      font-size: 16px;
-      margin-bottom: 4px;
-    }
-    
-    .stat-success {
-      color: #34a853;
-    }
-    
-    .stat-failure {
-      color: #ea4335;
-    }
-    
-    .stat-label {
-      color: #5f6368;
-      font-size: 12px;
-    }
-    
-    .loading-spinner {
-      width: 24px;
-      height: 24px;
-      border: 3px solid #e8eaed;
-      border-top: 3px solid #1a73e8;
-      border-radius: 50%;
-      animation: spin 1s linear infinite;
-      margin: 16px auto;
-    }
-    
-    @keyframes spin {
-      0% { transform: rotate(0deg); }
-      100% { transform: rotate(360deg); }
-    }
-    
-    .cancel-button {
-      background-color: #f8f9fa;
-      color: #3c4043;
-      border: 1px solid #dadce0;
-      padding: 8px 16px;
-      border-radius: 6px;
-      font-size: 13px;
-      cursor: pointer;
-      margin-top: 16px;
-      transition: background-color 0.2s ease;
-    }
-    
-    .cancel-button:hover {
-      background-color: #e8eaed;
-    }
-    
-    .note {
-      font-size: 12px;
-      color: #5f6368;
-      margin-top: 16px;
-      font-style: italic;
-    }
-  </style>
-</head>
-<body>
-  <div class="container">
-    <div class="title">
-      <span class="icon">🔍</span>
-      <?= title || 'URLFinder' ?> 実行中...
-    </div>
-    
-    <div class="message">
-      <?= message || '処理中...' ?>
-    </div>
-    
-    <div class="progress-container">
-      <div class="progress-bar">
-        <div class="progress-fill" id="progressFill" style="width: 0%"></div>
-      </div>
-      <div class="progress-text" id="progressText">
-        準備中...
-      </div>
-    </div>
-    
-    <div class="loading-spinner"></div>
-    
-    <div class="stats">
-      <div class="stat-item">
-        <span class="stat-value stat-success" id="successCount">0</span>
-        <span class="stat-label">成功</span>
-      </div>
-      <div class="stat-item">
-        <span class="stat-value stat-failure" id="failureCount">0</span>
-        <span class="stat-label">失敗</span>
-      </div>
-    </div>
-    
-    <button class="cancel-button" onclick="cancelExecution()">
-      キャンセル
-    </button>
-    
-    <div class="note">
-      処理中はブラウザを閉じないでください
-    </div>
-  </div>
-
-  <script>
-    let isCancelled = false;
-    
-    // プログレス更新関数
-    function updateProgress(current, total, success, failure) {
-      const percentage = total > 0 ? Math.round((current / total) * 100) : 0;
-      
-      // プログレスバーを更新
-      document.getElementById('progressFill').style.width = percentage + '%';
-      
-      // テキストを更新
-      document.getElementById('progressText').textContent = 
-        `${current}/${total}件完了 (${percentage}%)`;
-      
-      // 統計を更新
-      document.getElementById('successCount').textContent = success || 0;
-      document.getElementById('failureCount').textContent = failure || 0;
-    }
-    
-    // 完了時の表示更新
-    function showCompletion(stats) {
-      document.querySelector('.title').innerHTML = 
-        '<span class="icon">🎉</span>処理完了!';
-      
-      document.querySelector('.message').textContent = 
-        '全ての処理が完了しました';
-      
-      document.getElementById('progressFill').style.width = '100%';
-      document.getElementById('progressText').textContent = '完了';
-      
-      // スピナーを非表示
-      document.querySelector('.loading-spinner').style.display = 'none';
-      
-      // キャンセルボタンをOKボタンに変更
-      const cancelBtn = document.querySelector('.cancel-button');
-      cancelBtn.textContent = 'OK';
-      cancelBtn.onclick = () => google.script.host.close();
-      
-      // 統計更新
-      if (stats) {
-        document.getElementById('successCount').textContent = stats.success || 0;
-        document.getElementById('failureCount').textContent = stats.failure || 0;
-      }
-    }
-    
-    // 処理キャンセル
-    function cancelExecution() {
-      if (confirm('処理を中止しますか？')) {
-        isCancelled = true;
-        document.querySelector('.message').textContent = '処理を中止しています...';
-        document.querySelector('.cancel-button').disabled = true;
-        
-        // サーバーサイドのキャンセル処理を呼び出し（実装時に追加）
-        // google.script.run.cancelUrlFinderExecution();
-        
-        // 一旦、ダイアログを閉じる
-        setTimeout(() => {
-          google.script.host.close();
-        }, 1000);
-      }
-    }
-    
-    // エラー時の表示
-    function showError(message) {
-      document.querySelector('.title').innerHTML = 
-        '<span class="icon">⚠️</span>エラー発生';
-      
-      document.querySelector('.message').textContent = 
-        message || 'エラーが発生しました';
-      
-      // スピナーを非表示
-      document.querySelector('.loading-spinner').style.display = 'none';
-      
-      // キャンセルボタンを閉じるボタンに変更
-      const cancelBtn = document.querySelector('.cancel-button');
-      cancelBtn.textContent = '閉じる';
-      cancelBtn.onclick = () => google.script.host.close();
-    }
-    
-    // 初期化
-    document.addEventListener('DOMContentLoaded', function() {
-      // 初期プログレス設定
-      updateProgress(0, 1, 0, 0);
-    });
-  </script>
-</body>
-</html>
